requirements.txt
Django==5.0.3
djangorestframework==3.15.1
psycopg2-binary==2.9.9
django-environ==0.11.2
gunicorn==21.2.0
python-dotenv==1.0.1
